Retrieve Data from the List of the Mammals of Colombia

This function takes a list of Mammal species names, searches for their data in the MammalCol pacakage dataset, and returns a data frame containing the relevant information for each species.

Usage

search_mammalcol(splist, max_distance = 0.2)

Arguments

splist

A character vector containing the names of the species to search for.

max_distance

The maximum allowed distance for fuzzy matching of species names. Defaults to 0.2.

Value

A data frame containing the retrieved information for each species.

Details

The function allows fuzzy matching for species names with a maximum distance threshold to handle potential typos or variations in species names.

Examples

splist <- c("Tapirus bairdii", "Tapirus pinchaque", "Tapirus terrestris",
            "Tapirus terrestris", "tapir terrestre", "Pudu mephistophiles")

search_mammalcol(splist)
#> The following names are repeated in the 'splist': Tapirus terrestris
#>        name_submitted  kingdom   phylum    class          order    family
#> 1     Tapirus bairdii Animalia Chordata Mammalia Perissodactyla Tapiridae
#> 2   Tapirus pinchaque Animalia Chordata Mammalia Perissodactyla Tapiridae
#> 3  Tapirus terrestris Animalia Chordata Mammalia Perissodactyla Tapiridae
#> 5 Pudu mephistophiles Animalia Chordata Mammalia   Artiodactyla  Cervidae
#>     genus specificEpithet taxonRank scientificNameAuthorship
#> 1 Tapirus         bairdii   Especie             (Gill, 1865)
#> 2 Tapirus       pinchaque   Especie           (Roulin, 1829)
#> 3 Tapirus      terrestris   Especie         (Linnaeus, 1758)
#> 5    Pudu  mephistophiles   Especie        (de Winton, 1896)
#>               taxonRemarks
#> 1    Elevación (m): 0-1000
#> 2 Elevación (m): 1400-4000
#> 3    Elevación (m): 0-2400
#> 5 Elevación (m): 3000-3500
#>                                                                                                                                                                                                                                                           bibliographicCitation
#> 1                                                              HERSHKOVITZ P. 1954. Mammals of northern Colombia. Preliminary report No. 7: Tapirs (genus Tapirus), with a systematic review of American species. Proceedings of the United States National Museum 103:465-496.
#> 2                                                              ARIAS-ALZATE A, CC DOWNER, C DELGADO-V y J SÁNCHEZ-LONDOÑO. 2010. Un registro de tapir de montaña (Tapirus pinchaque) en el norte de la Cordillera Occidental de Colombia. Mastozoología Neotropical 17:111-116.
#> 3 ARIAS-ALZATE A, JA PALACIO VIEIRA y J MUÑOZ-DURAN. 2009. Nuevos registros de distribución y oferta de hábitat de la danta colombiana (Tapirus terrestris colombianus) en las tierras bajas del norte de la Cordillera Central (Colombia). Mastozoología Neotropical 16:19-25.
#> 5                                                                                                                                        ALBERICO M, A CADENA, JI HERNÁNDEZ-CAMACHO y Y MUÑOZ-SABA. 2000b. Mamíferos (Synapsida: Theria) de Colombia. Biota Colombiana 1:43-75.
#>   inMDD                                 Col_redlist
#> 1     1 EN A2cd+ B1ab (i,ii,iii)+ B2ab (i, ii, iii)
#> 2     1                                     EN A4cd
#> 3     1                                     VU A4cd
#> 5     1       EN A2cd B1ab (i; ii; iii; iv; v) \r\n
#>                                                                                                                                                                           distribution
#> 1                                                                                                                                                          Antioquia | Chocó | Córdoba
#> 2                                              Antioquia | Cauca | Cundinamarca | Huila | Meta | Nariño | Quindío | Risaralda | Tolima | Valle del Cauca | Caldas | Caquetá | Putumayo
#> 3 Antioquia | Amazonas | Arauca | Caquetá | Cauca | Cesar | Córdoba | Guainia | La Guajira | Meta | Magdalena | Vaupés | Vichada | Guaviare | Casanare | Norte de Santander | Putumayo
#> 5                                                                                                              Cauca | Huila | Nariño | Risaralda | Quindío | Tolima | Valle del Cauca
#>                                                         source endemic
#> 1 Proceedings of the United States National Museum 103:465-496      No
#> 2                         Mastozoología Neotropical 17:111-116      No
#> 3                           Mastozoología Neotropical 16:19-25      No
#> 5                                     Biota Colombiana 1:43-75      No
#>             english_name Distance
#> 1 Central American Tapir       12
#> 2         Mountain Tapir       14
#> 3          Lowland Tapir       16
#> 5          Northern Pudu       17