Mammal occurrence by departamento

Find the mammal species that occur in a given departamento of Colombia

sp_by_depto(states, type = c("any", "only", "all"), taxa = NULL)

Arguments

states

a character vector with one or more departamento names

type

type of matching to be used. any will return the mammal species that occur in any of the passed states. only matches mammal species that occur only in all provided (no more, no less) states and all matches mammal species that occur at least in all states passed. See examples.

taxa

optional character vector to match against the departamentos. Use the order name

Value

a data frame

Examples

library (mammalcol)

occ.any <- sp_by_depto(c("Arauca", "Norte de Santander"), type = "any")
occ.only <- sp_by_depto(c("Norte de Santander"), type = "only")
occ.all <- sp_by_depto(c("Arauca", "Norte de Santander"), type = "all")
occ.taxa <- sp_by_depto(c("Arauca", "Norte de Santander"), type = "all", taxa = "Chiroptera")

head(occ.any)
#>          scientificName         family        order
#> 1   Vampyressa voragine Phyllostomidae   Chiroptera
#> 2   Odocoileus goudotii       Cervidae Artiodactyla
#> 3       Chilomys fumeus     Cricetidae     Rodentia
#> 4 Melanomys columbianus     Cricetidae     Rodentia
#> 5        Tonatia maresi Phyllostomidae   Chiroptera
#> 6     Coendou pruinosus Erethizontidae     Rodentia
#>                                                                                         locality
#> 1                                                                  Casanare | Norte de Santander
#> 2                                         Boyacá | Cundinamarca | Santander | Norte de Santander
#> 3                                                                             Norte de Santander
#> 4                                                                 Magdalena | Norte de Santander
#> 5 Amazonas | Caquetá | Casanare | Huila | Meta | Putumayo | Vaupés | Guaviare | Arauca | Guainía
#> 6                Cundinamarca | Norte de Santander | Santander | Guania | Meta | Arauca | Tolima
head(occ.only)
#>    scientificName     family    order           locality
#> 1 Chilomys fumeus Cricetidae Rodentia Norte de Santander
head(occ.all)
#>               scientificName         family      order
#> 1 Syntheosciurus granatensis      Sciuridae   Rodentia
#> 2         Proechimys guairae     Echimyidae   Rodentia
#> 3         Alouatta seniculus       Atelidae   Primates
#> 4         Dermanura phaeotis Phyllostomidae Chiroptera
#> 5              Puma concolor        Felidae  Carnivora
#> 6           Leopardus wiedii        Felidae  Carnivora
#>                                                                                                                                                                                                                                                                                                locality
#> 1 Caldas | Cauca | Nariño | Quindío | Risaralda | Putumayo | Amazonas | Antioquia | Arauca | Atlántico | Bolívar | Boyacá | Casanare | Cesar | Chocó | Córdoba | Cundinamarca | Guaviare | Huila | La Guajira | Magdalena | Norte de Santander | Santander | Sucre | Tolima | Valle del Cauca | Vichada
#> 2                                                                                                                                                                                                                                                                           Arauca | Norte de Santander
#> 3                          Bolívar | Caldas | Cauca | Quindío | Risaralda | Valle del Cauca | Amazonas | Caquetá | Guaviare | Guainia | Nariño | Meta | Putumayo | Vichada | Antioquia | Arauca | Atlántico | Casanare | Cesar | Córdoba | La Guajira | Magdalena | Norte de Santander | Sucre | Tolima
#> 4                                                                                                Antioquia | Cauca | Chocó | Nariño | Quindío | Risaralda | Valle del Cauca | Amazonas | Arauca | Caldas | Caquetá | Córdoba | Guainía | Magdalena | Norte de Santander | Putumayo | Santander | Tolima
#> 5    Antioquia | Arauca | Bolívar | Córdoba | Caldas | Cauca | Caquetá | Casanare | Chocó | Magdalena | Meta | Nariño | Norte de Santander | Putumayo | Quindío | Risaralda | Sucre | Vaupés | Vichada | Amazonas | Guaviare | Atlántico | Cesar | Córdoba | Guainía | Huila | Tolima | Valle del Cauca
#> 6                                            Cauca | Quindío | Amazonas | Guainía | Meta | Putumayo | Antioquia | Arauca | Atlántico | Bolívar | Caldas | Caquetá | Casanare | Chocó | Córdoba | Cundinamarca | Guaviare | Magdalena | Nariño | Norte de Santander | Tolima | Valle del Cauca | Vichada
head(occ.taxa)
#>           scientificName         family      order
#> 1     Dermanura phaeotis Phyllostomidae Chiroptera
#> 2         Molossus rufus     Molossidae Chiroptera
#> 3  Cynomops planirostris     Molossidae Chiroptera
#> 4      Vampyressa thyone Phyllostomidae Chiroptera
#> 5    Enchisthenes hartii Phyllostomidae Chiroptera
#> 6 Micronycteris microtis Phyllostomidae Chiroptera
#>                                                                                                                                                                                                                                             locality
#> 1                                             Antioquia | Cauca | Chocó | Nariño | Quindío | Risaralda | Valle del Cauca | Amazonas | Arauca | Caldas | Caquetá | Córdoba | Guainía | Magdalena | Norte de Santander | Putumayo | Santander | Tolima
#> 2                                                                                             Arauca | Caldas | Norte de Santander | Tolima | Putumayo | Guaviare | Amazonas | Antioquia | Atlántico | Caquetá | Casanare | Guainía | Meta | Vichada
#> 3                                                                                                                  Amazonas | Arauca | Boyacá | Casanare | Meta | Vichada | Caquetá | Cundinamarca | Norte de Santander | Putumayo | Guainía | Sucre
#> 4 Caldas | Cauca | Quindío | Risaralda | Amazonas | Caquetá | Guaviare | Meta | Putumayo | Antioquia | Arauca | Casanare | Chocó | Córdoba | Guainía | Magdalena | Nariño | Norte de Santander | Sucre | Tolima | Valle del Cauca | Vaupés | Vichada
#> 5              Caldas | Cauca | Cundinamarca | Quindío | Risaralda | Amazonas | Caqueta | Putumayo | Antioquia | Arauca | Casanare | Cesar | Chocó | Guainía | Huila | Magdalena | Nariño | Norte de Santander | Putumayo | Tolima | Valle del Cauca
#> 6                                                                                                             Cundinamarca | Magdalena | Amazonas | Caquetá | Guaviare | Arauca | Casanare | Chocó | Cordoba | Guainia | Norte de Santander | Tolima