taxon Dataset

The taxon dataset comprises a tibble containing information regarding mammal species documented in Colombia, sourced from the "List of the mammals of Colombia" compiled by Ramírez-Chaves 2024. It encompasses various details, including scientific names, English names, Spanish names, order, family, threat status and distribution of each species. The table is a variant of the taxon table from SiB Colombia

taxon

Format

A tibble with 548 rows and 19 columns:

id

id from original taxon table.

scientificName

Scientific name of the mammal species.

kingdom

kingdom name of the mammal species.

phylum

phylum name of the mammal species.

class

class name of the mammal species.

order

order name of the mammal species.

family

family name of the mammal species.

genus

genus name of the mammal species.

specificEpithet

specificEpithet name of the mammal species.

taxonRank

taxon Rank name of the mammal species.

scientificNameAuthorship

species name´s author of the mammal species.

taxonRemarks

elevation of the mammal species.

bibliographicCitation

bibliographicCitation of the mammal species.

inMDD

1 if it is included on the Mammal Diverity Data Base, 0 not included.

Col_redlist

conservation status in Colombia for the mammal species.

distribution

Departamento were the mammal is present.

source

reference for the distribution.

endemic

Yes if it is endemic from Colombia, otherwise No.

english_name

english_name of the mammal species.

Details

This dataset is designed to provide users with comprehensive information about the mammal species found in Colombia, as documented by Ramírez-Chaves, et al. 2024. It is organized for easy access and utilization within the R environment.

References

The dataset is based on the "List of the Mammals of Colombia" by: Ramírez-Chaves H E, Leuro Robles N G, Castaño Rivera A, Morales-Martínez D M, Suárez Castro A F, Rodríguez-Posada M E, Zurc D, Concha Osbahr D C, Trujillo A, Noguera Urbano E A, Pantoja Peña G E, González Maya J F, Pérez Torres J, Mantilla Meluk H, López Castañeda C, Velásquez Valencia A, Zárrate Charry D (2024): Mamíferos de Colombia. v1.14. Sociedad Colombiana de Mastozoología. Dataset/Checklist SiB Colombia

See also

Author

Data compilation: Ramírez-Chaves 2021, Package implementation: Diego J. Lizcano

Examples

# Load the mammalcol package
library(mammalcol)

# Access the mammalcol_tab dataset
data("taxon")

# Display the first few rows
head(taxon)
#>             scientificName  kingdom   phylum    class        order   family
#> 1       Passalites murelia Animalia Chordata Mammalia Artiodactyla Cervidae
#> 2            Mazama rufina Animalia Chordata Mammalia Artiodactyla Cervidae
#> 3 Passalites sanctaemartae Animalia Chordata Mammalia Artiodactyla Cervidae
#> 4            Mazama temama Animalia Chordata Mammalia Artiodactyla Cervidae
#> 5            Mazama zamora Animalia Chordata Mammalia Artiodactyla Cervidae
#> 6             Mazama zetta Animalia Chordata Mammalia Artiodactyla Cervidae
#>        genus specificEpithet taxonRank scientificNameAuthorship
#> 1 Passalites         murelia   Especie        J. A. Allen, 1915
#> 2     Mazama          rufina   Especie         (Pucheran, 1851)
#> 3 Passalites   sanctaemartae   Especie      (J. A. Allen, 1915)
#> 4     Mazama          temama   Especie             (Kerr, 1792)
#> 5     Mazama          zamora   Especie        J. A. Allen, 1915
#> 6     Mazama           zetta   Especie             Thomas, 1913
#>               taxonRemarks
#> 1     Elevación (m): 0-900
#> 2 Elevación (m): 1500-4000
#> 3     Elevación (m): 0-900
#> 4      Elevación (m): 0-68
#> 5  Elevación (m): 600-1200
#> 6 Elevación (m): 1600-1950
#>                                                                                                                                                                    bibliographicCitation
#> 1 ALLEN JA. 1916. List of mammals collected in Colombia by the American Museum of Natural History expeditions, 1910-1915. Bulletin of the American Museum of Natural History 35:191-238.
#> 2                                                 ALBERICO M, A CADENA, JI HERNÁNDEZ-CAMACHO y Y MUÑOZ-SABA. 2000b. Mamíferos (Synapsida: Theria) de Colombia. Biota Colombiana 1:43-75.
#> 3 ALLEN JA. 1916. List of mammals collected in Colombia by the American Museum of Natural History expeditions, 1910-1915. Bulletin of the American Museum of Natural History 35:191-238.
#> 4  GROVES C y P GRUBB. 1987. Relationships of living deer. Pp. 21-59, en: Biology and management of the Cervidae (CM Wemmer, ed.). Smithsonian Institution Press, Washington, D.C., USA.
#> 5 ALLEN JA. 1916. List of mammals collected in Colombia by the American Museum of Natural History expeditions, 1910-1915. Bulletin of the American Museum of Natural History 35:191-238.
#> 6 ALLEN JA. 1916. List of mammals collected in Colombia by the American Museum of Natural History expeditions, 1910-1915. Bulletin of the American Museum of Natural History 35:191-238.
#>   inMDD   Col_redlist
#> 1     0          <NA>
#> 2     1 VU A2 cde\r\n
#> 3     0          <NA>
#> 4     1            DD
#> 5     0            DD
#> 6     0            DD
#>                                                                                                                                                              distribution
#> 1                                                                                 Amazonas | Caquetá | Cauca | Meta | Vaupés | Vichada | Nariño | Putumayo | Cundinamarca
#> 2 Antioquia | Boyacá | Caldas | Cauca | Cesar | Cundinamarca | Huila | Nariño | Norte de Santander | Quindío | Risaralda | Santander | Tolima | Valle del Cauca | Caquetá
#> 3                                                                                           Bolívar | Cesar | La Guajira | Magdalena | Atlántico | Córdoba | Cundinamarca
#> 4                                                                                                                                                       Chocó | Magdalena
#> 5                                                                                                                          Amazonas | Caquetá | Meta | Putumayo | Vichada
#> 6                                                                                                          Antioquia | Caldas | Cauca | Nariño | Valle del Cauca | Tolima
#>                                                                                                                         source
#> 1                                                                Bulletin of the American Museum of Natural History 35:191-238
#> 2                                                                                                     Biota Colombiana 1:43-75
#> 3                                                                Bulletin of the American Museum of Natural History 35:191-238
#> 4 Pp. 21-59, en: Biology and management of the Cervidae (CM Wemmer, ed.). Smithsonian Institution Press, Washington, D.C., USA
#> 5                                                                Bulletin of the American Museum of Natural History 35:191-238
#> 6                                                                Bulletin of the American Museum of Natural History 35:191-238
#>   endemic        english_name
#> 1      No                <NA>
#> 2      No  Little Red Brocket
#> 3      No                <NA>
#> 4      No Mexican Red Brocket
#> 5      No                <NA>
#> 6      No                <NA>
#>                                                                                    id
#> 1                                             urn:lsid:organismnames.com:name:3484299
#> 2 urn:lsid:catalogueoflife.org:taxon:4f16759f-5e17-11e7-8cee-bc764e092680:col20170824
#> 3                                                eol.org/pages/Mazama%20sanctaemartae
#> 4 urn:lsid:catalogueoflife.org:taxon:259114ac-5e17-11e7-8cee-bc764e092680:col20170824
#> 5                                             urn:lsid:organismnames.com:name:3484302
#> 6                                             urn:lsid:organismnames.com:name:3535811